overhaul Java classes
Since some people are actually looking at this repository, it might be a good idea to clean it up a bit.

This change makes the variable names more clear, changes the style of the comments, and, most importantly, changes the way the lookahead point is calculated to be more precise, since it doesn't use trigonometric functions anymore.